Uses the Finnegan mini-lib to test
Finnegan is a CC0-licensed single-file lib that uses regular expressions
as part of its validation of random text to ensure it doesn't include
unpronounceable or vulgar words. With the slight tweaks that this commit
makes, RegExodus can be dropped in as a substitute for java.util.Pattern
in Finnegan, and is likely to be a good substitute in a lot more code
too.