Optimizations. Removed unused filesystem code.
generator.js and thus Category.java can now identify valid Java
identifier parts, which replaces one GWT-incompatible method with one
that should be compatible. Another is replaced with a check for the
Unicode Z property to find any whitespace, since Character.isWhitespace
is missing on GWT (the fix may need tweaking). I ran the IntelliJ IDEA
inspections on this, and it fixed some things and may have made some
stupid changes in others (some methods that should be public may
currently be private; this will be fixed as I find them). It looks like
the derivative package will be moved to the non-compiled etc folder in
the next commit, since we have a working engine now and the speed is
decent.