Release 0.1.3 with fix to case-insensitive matching
A bug in case insensitive mode made capital letters unable to be found
when matching with the "i" flag on. This has been fixed and the current
behavior seems correct, though there may be a very small performance
impact in all modes. Now, with case insensitive mode on, when searching
for "([a-f])\\1" and replacing with "$1$1$1", "aaQ" becomes "aaaQ",
"BBQ" becomes "BBBQ", and so on.