Immensely frustrating bug fixed. 0.1.4
This is release 0.1.4, fixing an extremely hard-to-track-down bug
regarding a detail of the internal storage of character classes. The bug
meant certain chars, like 'û', would never be considered "word"
characters for the purposes of boundaries, and this could make certain
regexes fail or succeed in strange ways based solely on how close a
codepoint was to the end of a 256-codepoint block. 'û' was affected
because it was part of a span of valid 'word' characters that started in
the first block of 256 and ended several blocks later, and the character
that the span ended on happened to make the span have 0 length (other
lengths could have been possible in other cases). This only affects, as
far as I am aware, the Unicode-based '\\w', '\\W', '\\b', and '\\B'
escapes, for (possibly non-)word or boundary matching. This fix may
improve other Unicode boundary or category bugs.