Features added to backreferences
You can now specify whether you want a backreference that uses the
named-group syntax (which also works with numbered groups) to use
case-insensitive search with `@`, reversed-string search with `/`, or
matching-bracket search with `:`. For example
`Pattern.compile("([\\(\\[\\{]+){\\:/1}")` will match `()`, `[]`, and
`{}`, but not `{]` or `(}`. Because of the reverse search enabled by the
slash, `([])` will also match.