New backref features apply in replacements
You can now use modes in the named-group-referencing syntax, like
`${/1}`, in a replacement string. That example would resolve to the
string matched by group 1, but reversed because of the slash meaning
reverse mode. `@` for making the replacement lower-case and `:` for
mirroring brackets to and from opening and closing also can be added to
or replace `:` before the name or number of a group.