Little fix for debugging-OOME Heisenbug
Calling toString on some Term objects (which never are exposed outside
of this lib, but the call may happen when toString is called on
something visible) could have resulted in OutOfMemoryErrors, and this
frequently caused serious slowdowns when stepping through code to debug
other issues. This insidious bug, and its parent bug in IntBitSet, has
now been fixed.