Minor 0.1.7 release; fixes toString/serializing bug
This release is very small compared to the previous 0.1.6, but is
important if you expect to convert a Pattern to String; it often would
overflow the stack before. A specialized pair of methods allows you to
take a Pattern and serializeToString(), then later
deserializeFromString() to reproduce the Pattern, including flags (which
weren't tracked after construction before).