Some quick fixes to CharCharMap.
The copy constructor didn't copy the zeroValue or hasZeroValue fields. The place() method doesn't need long math because it always deals with numbers smaller than the precise limit on GWT, even with int multiplication. The hashCode() method can avoid the LEA issue by not multiplying by 31, and Entry.hashCode() was a total mess.